// apps/main/src/models/v2/controlplane/environmentPage.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { Environment } from './environment';

/**
 * A page of environments returned from the list api.
 */
export interface EnvironmentPage {
  /** The items in this page */
  items: Environment[];
  /** The page token to use to request the next page of items */
  next_page_token?: string;
}