// apps/main/src/models/v2/controlplane/environmentStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

/**
 * The status of the environment. Environments are normally active unless they have been deleted. A delete_failed status indicates that the destroy failed and a delete can be re-issued.
 */
export type EnvironmentStatus = (typeof EnvironmentStatus)[keyof typeof EnvironmentStatus];

export const EnvironmentStatus = {
  active: 'active',
  deleting: 'deleting',
  delete_failed: 'delete_failed',
} as const;